+++ components/header.js
@@ -1,10 +1,6 @@
-import { useRouter } from 'next/router'
 import ActiveLink from './ActiveLink'
 
 export default function Header() {
-  const router = useRouter();
-
-
   return (
     <header>
       <div>
