+++ pages/_app.js
@@ -1,9 +1,13 @@
 import '../public/index.css'
 import 'bootstrap/dist/css/bootstrap.min.css'
+import NavBar from '../components/navBar';
 
 
 
 export default function App({ Component, pageProps }) {
-  return <Component {...pageProps} />
+  return <>
+  <NavBar />
+  <Component {...pageProps} />
+  </>
 }
