+++ pages/index.js
@@ -2,7 +2,9 @@ import { initFirebase } from '../lib/firebaseApp';
 import { getAuth, signInWithPopup, GoogleAuthProvider } from "firebase/auth";
 import { useAuthState } from 'react-firebase-hooks/auth'
 import { useRouter } from 'next/router';
-
+import React, { useState } from 'react';
+import { Button, Form } from 'react-bootstrap';
+import Login from './login';
 
 export default function HomePage() {
   //initializing configuration
@@ -11,6 +13,16 @@ export default function HomePage() {
   const auth = getAuth();
   const [user, loading] = useAuthState(auth);
   const router = useRouter();
+
+  const [data, setData] = useState({
+    email: '',
+    password: '',
+  });
+
+  const handleLogin = async (e) => {
+    e.preventDefault();
+    console.log(data)
+  }
 
 
   if (loading) {
@@ -27,13 +39,60 @@ export default function HomePage() {
     const result = await signInWithPopup(auth, provider)
     console.log(result.user)
 
+  
+
+  
+
   }
   return (
     <div>
-      <h3>Click the button below to sign in:</h3>
-      <br></br>
-      <button onClick={() => signIn()}> Click me to if you want to see your dashboard!</button>
-      <br></br>
+       <div
+      style={{
+        width: '40%',
+        margin: 'auto',
+      }}
+    >
+      <h1 className="text-center my-3 ">Login</h1>
+      <Form onSubmit={handleLogin}>
+        <Form.Group className="mb-3" controlId="formBasicEmail">
+          <Form.Label>Email address</Form.Label>
+          <Form.Control
+            onChange={(e) =>
+              setData({
+                ...data,
+                email: e.target.value,
+              })
+            }
+            value={data.email}
+            required
+            type="email"
+            placeholder="Enter email"
+          />
+        </Form.Group>
+
+        <Form.Group className="mb-3" controlId="formBasicPassword">
+          <Form.Label>Password</Form.Label>
+          <Form.Control
+            onChange={(e) =>
+              setData({
+                ...data,
+                password: e.target.value,
+              })
+            }
+            value={data.password}
+            required
+            type="password"
+            placeholder="Password"
+          />
+        </Form.Group>
+        <Button variant="primary" type="submit">
+          Login
+        </Button>
+        <br></br>
+        <br></br>
+        <button onClick={() => signIn()}> I am the only one that works</button>
+      </Form>
+    </div>
     </div>
   );
 }
