+++ pages/login.js
@@ -1,63 +1,63 @@
-import React, { useState } from 'react';
-import { Button, Form } from 'react-bootstrap';
+// import React, { useState } from 'react';
+// import { Button, Form } from 'react-bootstrap';
 
-export default function Login() {
-  const [data, setData] = useState({
-    email: '',
-    password: '',
-  });
+// export default function Login() {
+//   const [data, setData] = useState({
+//     email: '',
+//     password: '',
+//   });
 
-  const handleLogin = async (e) => {
-    e.preventDefault();
-    console.log(data)
-  }
+//   const handleLogin = async (e) => {
+//     e.preventDefault();
+//     console.log(data)
+//   }
 
 
 
-  return (
-    <div
-      style={{
-        width: '40%',
-        margin: 'auto',
-      }}
-    >
-      <h1 className="text-center my-3 ">Login</h1>
-      <Form onSubmit={handleLogin}>
-        <Form.Group className="mb-3" controlId="formBasicEmail">
-          <Form.Label>Email address</Form.Label>
-          <Form.Control
-            onChange={(e) =>
-              setData({
-                ...data,
-                email: e.target.value,
-              })
-            }
-            value={data.email}
-            required
-            type="email"
-            placeholder="Enter email"
-          />
-        </Form.Group>
+//   return (
+//     <div
+//       style={{
+//         width: '40%',
+//         margin: 'auto',
+//       }}
+//     >
+//       <h1 className="text-center my-3 ">Login</h1>
+//       <Form onSubmit={handleLogin}>
+//         <Form.Group className="mb-3" controlId="formBasicEmail">
+//           <Form.Label>Email address</Form.Label>
+//           <Form.Control
+//             onChange={(e) =>
+//               setData({
+//                 ...data,
+//                 email: e.target.value,
+//               })
+//             }
+//             value={data.email}
+//             required
+//             type="email"
+//             placeholder="Enter email"
+//           />
+//         </Form.Group>
 
-        <Form.Group className="mb-3" controlId="formBasicPassword">
-          <Form.Label>Password</Form.Label>
-          <Form.Control
-            onChange={(e) =>
-              setData({
-                ...data,
-                password: e.target.value,
-              })
-            }
-            value={data.password}
-            required
-            type="password"
-            placeholder="Password"
-          />
-        </Form.Group>
-        <Button variant="primary" type="submit">
-          Login
-        </Button>
-      </Form>
-    </div>
-  );
-};
+//         <Form.Group className="mb-3" controlId="formBasicPassword">
+//           <Form.Label>Password</Form.Label>
+//           <Form.Control
+//             onChange={(e) =>
+//               setData({
+//                 ...data,
+//                 password: e.target.value,
+//               })
+//             }
+//             value={data.password}
+//             required
+//             type="password"
+//             placeholder="Password"
+//           />
+//         </Form.Group>
+//         <Button variant="primary" type="submit">
+//           Login
+//         </Button>
+//       </Form>
+//     </div>
+//   );
+// };
