+++ pages/profile/index.js
@@ -1,0 +1,7 @@
+export default function Profile() {
+  return (
+    <div>
+      <h1> Profile Page! </h1>
+    </div>
+  )
+}
