+++ pages/reports/index.js
@@ -1,10 +1,8 @@
-import Header from '../../components/header';
 
 export default function Reoprts() {
   return (
     <div>
-      <Header />
-      <h1> Here are your Reports </h1>
+      <h1> Reports Page! </h1>
     </div>
   )
 }
