+++ pages/settings/index.js
@@ -1,10 +1,7 @@
-import Header from '../../components/header';
-
 export default function Settings() {
   return (
     <div>
-      <Header />
-      <h1> Configure your setting here: </h1>
+      <h1> Setting Page! </h1>
     </div>
   )
 }
